backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.7.0
python-dateutil==2.9.0
